+++ js/src/forum/addBadgeListUserCard.js
